predictor/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
async-timeout==5.0.1
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
flatbuffers==24.12.23
gast==0.4.0
google-auth==2.37.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.69.0
grpcio-tools==1.50.0
h5py==3.12.1
idna==3.10
importlib_metadata==8.5.0
joblib==1.4.2
keras==3.5.0
Keras-Preprocessing==1.1.2
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
numpy==1.26.4
oauthlib==3.2.2
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
protobuf==4.25.5
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
scikit-learn==1.6.0
scipy==1.13.1
six==1.17.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.17.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.0
zipp==3.21.0
